-moved Flex, Grid and Toolbar styles + variables to teams theme
