--- packages/react/src/themes/base/componentStyles.ts
@@ -1,7 +1,0 @@
-export { default as Flex } from './components/Flex/flexStyles'
-export { default as FlexItem } from './components/Flex/flexItemStyles'
-
-export { default as Toolbar } from './components/Toolbar/toolbarStyles'
-export { default as ToolbarRadioGroup } from './components/Toolbar/toolbarRadioGroupStyles'
-
-export { default as Grid } from './components/Grid/gridStyles'
