--- packages/react/src/themes/base/componentVariables.ts
@@ -1,4 +1,0 @@
-export { default as Flex } from './components/Flex/flexVariables'
-export { default as FlexItem } from './components/Flex/flexItemVariables'
-
-export { default as Grid } from './components/Grid/gridVariables'
