--- packages/react/src/themes/base/index.ts
@@ -1,10 +1,0 @@
-import { ThemeInput } from '../types'
-import * as siteVariables from './siteVariables'
-import * as componentVariables from './componentVariables'
-import * as componentStyles from './componentStyles'
-
-export default {
-  siteVariables,
-  componentVariables,
-  componentStyles,
-} as ThemeInput
