--- packages/react/src/themes/base/siteVariables.ts
@@ -1,38 +1,0 @@
-//
-// COLORS
-//
-
-export { colors, contextualColors, emphasisColors, naturalColors, colorScheme } from './colors'
-
-//
-// FONT SIZES
-//
-export const fontSizes = {
-  smallest: '0.8rem',
-  smaller: '1rem',
-  small: '1.2rem',
-  medium: '1.4rem',
-  large: '1.8rem',
-  larger: '2.4rem',
-  largest: '3rem',
-}
-
-//
-// FONT WEIGHTS
-//
-export const fontWeightLight = 200
-export const fontWeightSemilight = 300
-export const fontWeightRegular = 400
-export const fontWeightSemibold = 600
-export const fontWeightBold = 700
-
-//
-// LINE HEIGHTS
-//
-export const lineHeightSmallest = 1.2
-export const lineHeightSmaller = 1.2
-export const lineHeightSmall = 1.3333
-export const lineHeightMedium = 1.4286
-export const lineHeightLarge = 1.3333
-export const lineHeightLarger = 1.3333
-export const lineHeightLargest = 1.3333
