+++ packages/react/src/themes/index.ts
@@ -1,5 +1,4 @@
 // Themes
-export { default as base } from './base'
 export { default as fontAwesome } from './font-awesome'
 export { default as teams } from './teams'
 export { default as teamsDark } from './teams-dark'
