+++ packages/react/src/themes/teams/componentStyles.ts
@@ -85,6 +85,8 @@ export { default as Status } from './components/Status/statusStyles'
 
 export { default as Text } from './components/Text/textStyles'
 
+export { default as Toolbar } from './components/Toolbar/toolbarStyles'
+export { default as ToolbarRadioGroup } from './components/Toolbar/toolbarRadioGroupStyles'
 export { default as ToolbarCustomItem } from './components/Toolbar/toolbarCustomItemStyles'
 export { default as ToolbarDivider } from './components/Toolbar/toolbarDividerStyles'
 export { default as ToolbarItem } from './components/Toolbar/toolbarItemStyles'
