+++ packages/react/src/themes/teams/components/Grid/gridStyles.ts
@@ -2,8 +2,36 @@ import { GridVariables } from '../../../base/components/Grid/gridVariables'
 import { ComponentSlotStylesInput, ICSSInJSStyle } from '../../../types'
 import { GridProps } from '../../../../components/Grid/Grid'
 
+const getCSSTemplateValue = (template: string | number): string => {
+  const templateAsNumber = Number(template)
+
+  return !isNaN(templateAsNumber) && templateAsNumber > 0
+    ? `repeat(${template}, 1fr)`
+    : String(template)
+}
+
 const gridStyles: ComponentSlotStylesInput<GridProps, GridVariables> = {
-  root: (): ICSSInJSStyle => ({ '& > *': { outlineOffset: '-3px' } }),
+  root: ({
+    props,
+    variables: { height, width, defaultColumnCount, gridGap, padding },
+  }): ICSSInJSStyle => {
+    const { rows, columns = !props.rows && defaultColumnCount } = props
+
+    return {
+      height,
+      width,
+      padding,
+      gridGap,
+      display: 'grid',
+      justifyContent: 'space-evenly',
+
+      ...(rows && !columns && { gridAutoFlow: 'column' }),
+      ...(rows && { gridTemplateRows: getCSSTemplateValue(rows) }),
+      ...(columns && { gridTemplateColumns: getCSSTemplateValue(columns) }),
+
+      '& > *': { outlineOffset: '-3px' },
+    }
+  },
 }
 
 export default gridStyles
